+++ cider-repl.el
@@ -1034,8 +1034,6 @@ ENDP) DELIM."
     (define-key map (kbd "C-c M-m") 'cider-macroexpand-all)
     (define-key map (kbd "C-c C-z") 'cider-switch-to-last-clojure-buffer)
     (define-key map (kbd "C-c M-s") 'cider-selector)
-    (define-key map (kbd "C-c M-r") 'cider-rotate-connection)
-    (define-key map (kbd "C-c M-d") 'cider-display-current-connection-info)
     (define-key map (kbd "C-c M-f") 'cider-load-fn-into-repl-buffer)
     (define-key map (kbd "C-c C-q") 'cider-quit)
     (define-key map (string cider-repl-shortcut-dispatch-char) 'cider-repl-handle-shortcut)